fix: formatting issues

+++ app/lib/util/color.dart
@@ -6,7 +6,7 @@ extension ColorBrightness on Color {
 
     final hsl = HSLColor.fromColor(this);
     final hslLight =
-        hsl.withLightness((hsl.lightness + amount).clamp(0.0, 1.0).toDouble());
+        hsl.withLightness((hsl.lightness + amount).clamp(0.0, 1.0));
 
     return hslLight.toColor();
   }
